RF:https://github.com/RedRoverSchool/JenkinsQA_2024_fall/issues/1159  testPipelineDisabledTooltipOnHomePage -> PipelineProjectTest,  delete PipelineTest
